Add files via upload
Context-
Using science and technology to forecast the atmospheric conditions at a certain area and time is known as weather forecasting. Weather forecasting has been done sporadically for millennia and formally since the 19th century. Meteorology is used to predict how the atmosphere will change at a certain location while quantitative data about the state of the atmosphere, land, and ocean are gathered to provide weather forecasts.
Weather forecasting currently relies on computer-based models that take many atmospheric parameters into account, as opposed to the manual calculations of the past that were mostly focused on changes in barometric pressure, the current weather, and the state of the sky or cloud cover. o choose the optimal forecast model to use as the foundation for the forecast, human input is still necessary. This requires pattern recognition abilities, teleconnections, model performance knowledge, and model bias awareness. Forecasting is inaccurate because of the atmosphere's chaotic character, the enormous computing power needed to solve the equations that describe the atmosphere, land, and ocean, the measurement error associated with initial circumstances, and a lack of knowledge about atmospheric and related processes. As a result, forecasts lose accuracy when the interval between the current moment and the time for which it is being made (the forecast's range) widens. The usage of ensembles and model consensus aids in reducing error and offers a forecast's level of confidence. Weather forecasts can be used for a huge range of purposes. Weather warnings are crucial forecasts because they serve to safeguard property and human life. For traders in the commodity markets, forecasts based on temperature and precipitation are crucial for agriculture. Utility companies use temperature projections to predict demand for the upcoming days. Many people often consult weather forecasts to decide what to dress each day. Heavy rain, snow, and wind chill significantly restrict outdoor activities, therefore forecasts can be used to schedule activities around these conditions as well as to prepare for and survive them. The economy includes weather forecasting; for instance, in 2009, the US spent around 5.1 billion dollars on it, with benefits projected to be worth six times as much.
Investigative Data Analysis
 
Goal: To comprehend the data ("A small step forward is better than a big one backwards")
• Start drafting a modelling plan.
Features
- station - used weather station number: 1 to 25
- Date - Present day: yyyy-mm-dd ('2013-06-30' to '2017-08-30')
- Present_Tmax - Maximum air temperature between 0 and 21 h on the present day (Â°C): 20 to 37.6
- Present_Tmin - Minimum air temperature between 0 and 21 h on the present day (Â°C): 11.3 to 29.9
- LDAPS_RHmin - LDAPS model forecast of next-day minimum relative humidity (%): 19.8 to 98.5
- LDAPS_RHmax - LDAPS model forecast of next-day maximum relative humidity (%): 58.9 to 100
- LDAPSTmaxlapse - LDAPS model forecast of next-day maximum air temperature applied lapse rate (Â°C): 17.6 to 38.5
- LDAPSTminlapse - LDAPS model forecast of next-day minimum air temperature applied lapse rate (Â°C): 14.3 to 29.6
- LDAPS_WS - LDAPS model forecast of next-day average wind speed (m/s): 2.9 to 21.9
- LDAPS_LH - LDAPS model forecast of next-day average latent heat flux (W/m2): -13.6 to 213.4
- LDAPS_CC1 - LDAPS model forecast of next-day 1st 6-hour split average cloud cover (0-5 h) (%): 0 to 0.97
- LDAPS_CC2 - LDAPS model forecast of next-day 2nd 6-hour split average cloud cover (6-11 h) (%): 0 to 0.97
- LDAPS_CC3 - LDAPS model forecast of next-day 3rd 6-hour split average cloud cover (12-17 h) (%): 0 to 0.98
- LDAPS_CC4 - LDAPS model forecast of next-day 4th 6-hour split average cloud cover (18-23 h) (%): 0 to 0.97
- LDAPS_PPT1 - LDAPS model forecast of next-day 1st 6-hour split average precipitation (0-5 h) (%): 0 to 23.7
- LDAPS_PPT2 - LDAPS model forecast of next-day 2nd 6-hour split average precipitation (6-11 h) (%): 0 to 21.6
- LDAPS_PPT3 - LDAPS model forecast of next-day 3rd 6-hour split average precipitation (12-17 h) (%): 0 to 15.8
- LDAPS_PPT4 - LDAPS model forecast of next-day 4th 6-hour split average precipitation (18-23 h) (%): 0 to 16.7
- lat - Latitude (Â°): 37.456 to 37.645
- lon - Longitude (Â°): 126.826 to 127.135
- DEM - Elevation (m): 12.4 to 212.3
- Slope - Slope (Â°): 0.1 to 5.2
- Solar radiation - Daily incoming solar radiation (wh/m2): 4329.5 to 5992.9
- Next_Tmax - The next-day maximum air temperature (Â°C): 17.4 to 38.9
- Next_Tmin - The next-day minimum air temperature (Â°C): 11.3 to 29.8T  
Base Chacklist
Shape analysis :
•	target(s) : Next_Tmax & Next_Tmin
•	rows and columns : 7752 , 25
•	features types : qualitatives : 1 (la date du relevé) , quantitatives : 24 (tout le reste)
•	NaN analysis :
	vraiment pas beaucoup de NaN (moitié des variables = 1% de NaN)
Features analysis :
•	Target visualization :
	Next_Tmax = 30.27° +/- 3.12°
	Next_Tmin = 22.930 +/- 2.49°
Conclusion-
By using RandomForest, we can reach accuracies on both targets up to 97% (+/-0.65°C), which is truely decent.